feat: added TensorCollection method templates (most likely will be chaged)
