+++ src/tensorbase/include/TensorBase/core/sources.cmake
@@ -6,6 +6,7 @@ set(sources_list_h
 	Helloworld.h
 	Statistics.h
 	StringParsing.h
+	TupleAlgorithms.h
 )
 
 ### add path to the filenames
